Workaround for semantic names in Dotty
There is a weird bug that took me several days to figure out.
Names are semantic in Dotty, it means two names "temp$2" and
"temp$2" are not necessarily equal, which is the cause of
not found identifier error.

Concretely, semantic names imply the following two names are not
equal:

    val name1 = NameKinds.UniqueName.fresh("temp".toTermName)
    val name2 = name1.toString.toTermName

    assert(name1 != name2)

We avoid the problem by always build semantic names by `name.toTermName`.